feat: update deps and refactor
