+++ .nvmrc
@@ -1,0 +1,1 @@
+18.16.0
