+++ src/App.js
@@ -1,8 +1,24 @@
 import React from 'react';
-import Layout from './layouts/layout';
+import { Routes, Route, BrowserRouter } from 'react-router-dom';
+import About from './pages/About';
+import Articles from './pages/Articles';
+import OpenSource from './pages/OpenSource';
+import Projects from './pages/Projects';
+import Layout from './layout/MainLayout';
 
-function App() {
-  return <Layout />;
-}
+export const App = () => {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Layout />}>
+          <Route path="/" element={<About />} />
+          <Route path="/articles" element={<Articles />} />
+          <Route path="/open-source" element={<OpenSource />} />
+          <Route path="/projects" element={<Projects />} />
+        </Route>
+      </Routes>
+    </BrowserRouter>
+  );
+};
 
 export default App;
